refactor: 프론트 연동관련 코드수정

+++ src/main/java/com/home/nanda/house/model/dto/Deal.java
@@ -4,14 +4,14 @@ import lombok.Data;
 
 @Data
 public class Deal {
-	int dealCode;
+	Long dealCode;
 	String dealPrice;
-	int dealYear;
-	int dealMonth;
-	int dealDay;
+	Long dealYear;
+	Long dealMonth;
+	Long dealDay;
 	String area;
 	String floor;
-	int houseCode;
+	Long houseCode;
 	String type;
 	String rentMoney;
 }
